Add one more route specification method
If we have access to an API-based trip planner, for the region of interest,
we can call it manually for any kind of trip and then put the resulting
polyline directly into the spec file. In this case, the conversion is trivial,
as all we have to do is to decode the polyline.

TODO: Should we just use polylines for the specs? Will certainly make them be
shorter.

Also support using CAR to get BUS routes, since OSM relation data for bus
routes is spotty.
https://gis.stackexchange.com/questions/329249/how-to-convert-an-osm-relation-to-a-list-of-lat-lng-points

Barring dedicated bus lanes or BRT, buses travel on regular
roads, so we can potentially use driving directions to find the appropriate
route.

+ handle the two new modes (relation and polyline) in the validation since they
don't have associated waypoints
+ add a new notebook to make it easier to create ground truth for the various
legs interactively