Change the ranges to be based on the write timestamps
instead of transition timestamps.

This is needed because the data timestamps are generated in javascript first
and then the write timestamps are generated in the native code when the entry
is saved. Most of the time, this slight difference does not matter, but in our
case, we save a lot of entries in quick succession from code, *and* some of the
entries (battery) are saved completely in native code, so they don't have this
mismatch.

Without this fix, all the battery entries for a range were after the end
transition of the range, although our code stores the battery before generating
the transition.

As an example, consider the ranges for for `ucb-sdb-ios-1` in the first set of
calibration tests after the phone fixes. The stop transition's `ts` was
`2019-07-18T07:23:32.130000-07:00` while the write_ts was
`2019-07-18T07:23:32.221209-07:00`

But the last entry in the battery_df, the most meaningful one, was at
`2019-07-18T07:23:32.197008-07:00`. So after the `ts` but *before* the
`write_ts`. If we only use the `ts`, we only get the entries from the start of
the range, not the end, which makes the calibration useless.

```
START_CALIBRATION_PERIOD 2019-07-18T00:50:12.328000-07:00 2019-07-18T00:50:12.410647-07:00
STOP_CALIBRATION_PERIOD 2019-07-18T07:23:58.511000-07:00 2019-07-18T07:23:58.581305-07:00
23626.170658111572 2019-07-18T00:50:12.410647-07:00 2019-07-18T07:23:58.581305-07:00
   battery_level_pct                          fmt_time  \
0         100.000000  2019-07-18T00:49:57.993955-07:00
1         100.000000  2019-07-18T00:49:58.115479-07:00
2         100.000000  2019-07-18T00:49:58.354412-07:00
3         100.000000  2019-07-18T00:49:58.505404-07:00
4         100.000000  2019-07-18T00:49:58.587231-07:00
5         100.000000  2019-07-18T00:49:58.702553-07:00
6         100.000000  2019-07-18T00:49:58.880366-07:00
7          74.000001  2019-07-18T07:23:32.196434-07:00

                     write_fmt_time
0  2019-07-18T00:49:57.995355-07:00
1  2019-07-18T00:49:58.115692-07:00
2  2019-07-18T00:49:58.354609-07:00
3  2019-07-18T00:49:58.505546-07:00
4  2019-07-18T00:49:58.587383-07:00
5  2019-07-18T00:49:58.702698-07:00
6  2019-07-18T00:49:58.880602-07:00
7  2019-07-18T07:23:32.197008-07:00
```