Check in version that works for iOS
Fixed the parsing code so that there are no errors (see
19f7f09dfd734235a1cc9a218dd240eca1e41971,
7331881c71bc12994dbbed7864d78c38d86b149a)

Battery values work, location values work for iOS since that was the phone used
for ground truthing. Last step is to copy over the ranges from iOS to android
so that we can have the evaluation trips for android as well.

But let's get this intermediate step checked in to unblock others