Check in the filled timelines as well
We technically don't need to do this since:
- the filled specs are generated from the regular specs
- the filled specs are available on the server anyway

However, the process to go from regular to filled is not currently reproducible
(see sm_reroutes) since it does not take into account the OSM version at the
time of the generation.

And while it is true that the filled specs are on the server, it is also useful
to have them checked in as an example and to browse the data.