Partial fix for the notebook to validate the filled spec
This is partial fix to the notebook that allows us to visualize the filled spec
before we upload it.

The original version of the notebook still works with the original version of the filled spec
"final_sfbayarea_filled_reroutes/train_bus_ebike_mtv_ucb.filled.reroute.json"

I encourage you to run the original version of the notebook first, and
understand what it does. You might also want to save a local, unmodified copy
that you can refer to while fixing the notebook.

This PR attempts to change the notebook to visualize the new spec format.

Concretely, it renames the existing `get_marker` to `get_one_marker`
and adds a new `get_marker` function to call `get_one_marker` with each
location in the list.

It then modifies `get_map_for_travel_leg` to add all the markers to the map.
This means that we will have the time-dependent variations of the leg displayed
on the same map.

I then discovered a regression in the new `route_coords`.

The original version had the `route_coords` as a valid geojson object

```
          "route_coords": {
            "type": "Feature",
            "properties": {},
            "geometry": {
              "type": "LineString",
              "coordinates": [
                [
                  -122.08337,
                  37.39025
                ],
                [
                  -122.08338,
                  37.39022
                ],
```

The new one does not

```
          "route_coords": [
            [
              37.39025,
              -122.08337
            ],
            [
              37.39022,
              -122.08338
            ],
            [
              37.39015,
              -122.08337
            ],
```

This breaks the code where we try to access the `geometry` child of `route_coords`

```
len(trip["route_coords"]["geometry"]["coordinates"])
```

Please fix the regression and attempt to fix the rest of the visualization (the
route part) similar to the fix for the start and end locations.

Bonus points if you can have a popup displaying the validity show up when the
user clicks on any map feature.