Fix the duplicate checking
so that it doesn't report errors if all entries between the duplicates are also
duplicates.

Concretely, if entries 31,32 and 33 are all duplicates, the previous code would
have returned an error.  But it is not really an error, since there would not
be any zigzag, just repeated points. This checks the intermediate points to see
if they are identical as well by adding an addition `for` loop.