Improve the implementation of the eval_view
+ change the notebooks to match
+ switch everything to the filespec
+ bump up the python version on the final timeline

The original implementation had us pass in the start and end timestamp for the
ground truth to the viz function. But this could be problematic, as in the case
of the berkeley timeline, where there the ground truth depends on the start and
end timestamp.

So we get the ground truth for range that corresponds to the section that we
are trying to plot instead. This also simplifies the calling interface to the
function.