:camera_flash: Download one raw transition per program
After updating the specs to use E_BIKE instead of BICYCLING
https://github.com/MobilityNet/mobilitynet-analysis-scripts/commit/43bb437fb981801fee0f123a21d053f9bc95939f
https://github.com/MobilityNet/mobilitynet-analysis-scripts/commit/edbc4483b18194d5a6b70600fed54ca371a3ef3a
I reloaded the data into a local database using `spec_creation/upload_validated_spec.py`
and took a new mongodump snapshot

On re-downloading (`bin/dump_data_to_file.py`), the $oid and the write timestamps changed.
But we also got a new transition in the timeline.
I restored to the older database snapshot, and the transitions were still downloaded

This is likely due to the restructuring of the dump script, in the commits below
https://github.com/MobilityNet/mobilitynet-analysis-scripts/commit/be37dde875e0735cc940fcc9ff345c4b3c1c1bce
https://github.com/MobilityNet/mobilitynet-analysis-scripts/commit/e78da4f83c6d82cd417e49edccf81b95da11b93f
https://github.com/MobilityNet/mobilitynet-analysis-scripts/commit/dd6ba0f14c4e3893f51fce27c999fdc3726cf51e
https://github.com/MobilityNet/mobilitynet-analysis-scripts/commit/f4f48c3343c1a74c3821f5a14ceac3f4fef4a05b
I did not re-download raw data after the dump script changes.

I verified that the new transition did not change the outputs
Checking these in to make the csv files snapshot consistent with the mongodump snapshot