Fix the ground truth leg for `walk_to_bus`
On the last data collection run, I wasn't allowed to take the LBNL shuttle, and
had to take the AC transit bus up instead. And that was in a different start location.

This is a partial fix for
https://github.com/MobilityNet/mobilitynet.github.io/issues/25

For the fix to be complete, the change also needs to be incorporated into
the spec_creation folder and the saved MobilityNet mongodb